fix(core): properly print overrides for dot notation
